Console program to implement queues
A console program that simulates software for an "autobody" shop that would be used to control customer queues for vehicles that are waiting for service, or are currently in service. Vectors and a stack are used to record and save vehicle information for each customer, and also if their service bill has currently been paid or not. The purpose of this program was to develop skills to implement a vector, a queue, as well as to manipulate a stack