Typing tutor program in Windows Forms
This is a typing tutor program that was developed using windows forms. It generates a list of randomly selected falling letters and then draws them to a drawing canvas called GDIDrawer. The speed of the falling letter animation and the difficulty of the game is selected through the use of modal and modeless forms that are controlled by the main form. As the user plays each letter they type that corresponds to a current falling letter will add a point to their total score, remove that letter, and generate a new letter. The game is over when the user hits the game over button or a letter falls below the bottom of the screen. High scores for each difficulty level (easy, medium, hard) are saved in a serializable binary file, which is automatically generated if the program doesn't have one on startup.