Typing Tutor main form code
This is the code within the main cs file for the typing tutor program. This program focused on the use of multiple threads to control different processes while still maintaining the ability to alter contents in the main form, and through the use of delegates in order to exchange updated information between different windows forms.